+++ SwarmBots/main.py
@@ -1,0 +1,8 @@
+from SharedGrid import SharedGrid
+
+def main():
+    grid = SharedGrid(5, 5)
+    print(grid)
+
+if __name__ == '__main__':
+    main()
